Settled on public-read for s3 bucket for pretty URIs
I'd forgotten that in order to facilitate ignoring index.html files in
deep URIs, the cloudfront distribution requires that you use the s3
bucket's website endpoint instead of domain. As such, we can't attach an
IAM identity to the cloudfront distribution & then set it with the s3
origin config. Rather, we have to lease the s3 bucket as public-read.
Given that this module is for creating a public website, this seems like
a reasonable tradeoff.

This blogpost covers the basis issue:
https://stevepapa.com/how-to-specify-a-default-root-object-for-static-website-subdirectories-on-aws-cloudfront/